[WIP] Move and simplify TaskProvider
