+++ core/src/main/kotlin/at/bitfire/davdroid/settings/migration/AccountSettingsMigration19.kt
@@ -9,7 +9,7 @@ import android.content.Context
 import android.provider.CalendarContract
 import androidx.work.WorkManager
 import at.bitfire.davdroid.sync.AutomaticSyncManager
-import at.bitfire.ical4android.TaskProvider
+import at.bitfire.synctools.storage.TaskProvider
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -36,7 +36,7 @@ class AccountSettingsMigration19 @Inject constructor(
         val authorities = listOf(
             "at.bitfire.davdroid.addressbooks",
             CalendarContract.AUTHORITY,
-            *TaskProvider.TASK_PROVIDERS.map { it.authority }.toTypedArray()
+            TaskProvider.ProviderName.entries.map { it.authority }.toTypedArray()
         )
         for (authority in authorities) {
             val oldWorkerName = "periodic-sync $authority ${account.type}/${account.name}"
