+++ core/src/main/kotlin/at/bitfire/davdroid/ui/DebugInfoGenerator.kt
@@ -43,7 +43,7 @@ import at.bitfire.davdroid.sync.account.InvalidAccountException
 import at.bitfire.davdroid.sync.adapter.SyncFrameworkIntegration
 import at.bitfire.davdroid.sync.worker.BaseSyncWorker
 import at.bitfire.davdroid.util.TextTable
-import at.bitfire.ical4android.TaskProvider
+import at.bitfire.synctools.storage.TaskProvider
 import at.techbee.jtx.JtxContract
 import dagger.hilt.android.qualifiers.ApplicationContext
 import org.dmfs.tasks.contract.TaskContract
