+++ core/src/main/kotlin/at/bitfire/davdroid/ui/PermissionsScreen.kt
@@ -42,7 +42,7 @@ import at.bitfire.davdroid.ui.composable.AppTheme
 import at.bitfire.davdroid.ui.composable.CardWithImage
 import at.bitfire.davdroid.ui.composable.PermissionSwitchRow
 import at.bitfire.davdroid.util.PermissionUtils
-import at.bitfire.ical4android.TaskProvider
+import at.bitfire.synctools.storage.TaskProvider
 import java.util.logging.Level
 import java.util.logging.Logger
