+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/storage/jtx/JtxCollectionProviderTest.kt
@@ -8,7 +8,7 @@ import android.accounts.Account
 import android.content.ContentProviderClient
 import androidx.core.content.contentValuesOf
 import androidx.test.platform.app.InstrumentationRegistry
-import at.bitfire.ical4android.TaskProvider
+import at.bitfire.synctools.storage.TaskProvider
 import at.bitfire.synctools.test.GrantPermissionOrSkipRule
 import at.techbee.jtx.JtxContract
 import org.junit.After
