+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/storage/tasks/TestTaskList.kt
@@ -6,7 +6,7 @@ package at.bitfire.synctools.storage.tasks
 
 import android.accounts.Account
 import androidx.core.content.contentValuesOf
-import at.bitfire.ical4android.TaskProvider
+import at.bitfire.synctools.storage.TaskProvider
 import org.dmfs.tasks.contract.TaskContract
 
 object TestTaskList {
